requirements.txt
fastapi>=0.95.0,<0.110.0
uvicorn>=0.22.0,<0.28.0
python-dotenv>=1.0.0
supabase>=2.3.0,<3.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.2.0,<2.0.0
python-jose[cryptography]>=3.3.0
requests>=2.31.0
python-multipart>=0.0.9
openpyxl>=3.1.0
pydantic>=2.0.0,<3.0.0
asyncpg>=0.29.0
httpx>=0.24.0,<0.27.0 